@@ -1599,7 +1599,6 @@
         "@types/request": "^2.47.0",
         "@types/semver": "^5.5.0",
         "@types/shortid": "^0.0.29",
-        "@types/sinon": "^4.3.0",
         "@types/tmp": "0.0.33",
         "@types/untildify": "^3.0.0",
         "@types/uuid": "^3.4.3",
@@ -1637,12 +1636,12 @@
         "remap-istanbul": "^0.10.1",
         "retyped-diff-match-patch-tsd-ambient": "^1.0.0-0",
         "shortid": "^2.2.8",
-        "sinon": "^4.4.5",
+        "ts-mockito": "^2.3.1",
         "tslint": "^5.9.1",
         "tslint-eslint-rules": "^5.1.0",
         "tslint-microsoft-contrib": "^5.0.3",
         "typemoq": "^2.1.0",
-        "typescript": "^2.9.1",
+        "typescript": "^3.1.1",
         "typescript-formatter": "^7.1.0",
         "uuid": "^3.2.1",
         "vscode": "^1.1.21",
