@@ -8,7 +8,7 @@ import { inject, injectable, unmanaged } from 'inversify';
 import { IServiceContainer } from '../../ioc/types';
 import { captureTelemetry } from '../../telemetry';
 import { PYTHON_LANGUAGE_SERVER_LIST_BLOB_STORE_PACKAGES } from '../../telemetry/constants';
-import { log, LogOptions } from '../logger';
+import { LogOptions, traceVerbose } from '../logger';
 import { INugetRepository, INugetService, NugetPackage } from './types';
 
 @injectable()
@@ -21,7 +21,7 @@ export class AzureBlobStoreNugetRepository implements INugetRepository {
     }
 
     @captureTelemetry(PYTHON_LANGUAGE_SERVER_LIST_BLOB_STORE_PACKAGES)
-    @log('Listing Nuget Packages', LogOptions.Arguments)
+    @traceVerbose('Listing Nuget Packages')
     public listPackages(azureBlobStorageAccount: string, azureBlobStorageContainer: string, packageName: string) {
         const blobStore = createBlobServiceAnonymous(azureBlobStorageAccount);
         const nugetService = this.serviceContainer.get<INugetService>(INugetService);
