@@ -3,22 +3,21 @@ import { AttachRequestArguments, LaunchRequestArguments } from '../Common/Contra
 import { IDebugLauncherScriptProvider } from '../types';
 import { DebugClient } from './DebugClient';
 import { DebuggerLauncherScriptProvider, NoDebugLauncherScriptProvider } from './launcherProvider';
-import { LocalDebugClient } from './LocalDebugClient';
 import { LocalDebugClientV2 } from './localDebugClientV2';
 import { NonDebugClientV2 } from './nonDebugClientV2';
 import { RemoteDebugClient } from './RemoteDebugClient';
 
+type DebugClientArgs = [LaunchRequestArguments, DebugSession, boolean, IDebugLauncherScriptProvider];
 export function CreateLaunchDebugClient(launchRequestOptions: LaunchRequestArguments, debugSession: DebugSession, canLaunchTerminal: boolean): DebugClient<{}> {
     let launchScriptProvider: IDebugLauncherScriptProvider;
-    let debugClientClass: typeof LocalDebugClient;
+    const args: DebugClientArgs = [launchRequestOptions, debugSession, canLaunchTerminal, launchScriptProvider];
     if (launchRequestOptions.noDebug === true) {
         launchScriptProvider = new NoDebugLauncherScriptProvider();
-        debugClientClass = NonDebugClientV2;
+        return new NonDebugClientV2(...args);
     } else {
         launchScriptProvider = new DebuggerLauncherScriptProvider();
-        debugClientClass = LocalDebugClientV2;
+        return new LocalDebugClientV2(...args);
     }
-    return new debugClientClass(launchRequestOptions, debugSession, canLaunchTerminal, launchScriptProvider);
 }
 export function CreateAttachDebugClient(attachRequestOptions: AttachRequestArguments, debugSession: DebugSession): DebugClient<{}> {
     return new RemoteDebugClient(attachRequestOptions, debugSession);
