@@ -15,9 +15,9 @@ import { EXTENSION_ROOT_DIR } from '../../client/common/constants';
 import '../../client/common/extensions';
 import { IS_WINDOWS } from '../../client/common/platform/constants';
 import { IPlatformService } from '../../client/common/platform/types';
-import { PythonV2DebugConfigurationProvider } from '../../client/debugger';
 import { DebuggerTypeName, PTVSD_PATH } from '../../client/debugger/Common/constants';
 import { AttachRequestArguments, DebugOptions } from '../../client/debugger/Common/Contracts';
+import { PythonV2DebugConfigurationProvider } from '../../client/debugger/extension/configProviders/pythonV2Provider';
 import { IServiceContainer } from '../../client/ioc/types';
 import { PYTHON_PATH, sleep } from '../common';
 import { IS_MULTI_ROOT_TEST, TEST_DEBUGGER } from '../initialize';
