@@ -14,12 +14,11 @@ import { PYTHON_LANGUAGE } from '../../../client/common/constants';
 import { IFileSystem, IPlatformService } from '../../../client/common/platform/types';
 import { IPythonExecutionFactory, IPythonExecutionService } from '../../../client/common/process/types';
 import { IConfigurationService, ILogger, IPythonSettings } from '../../../client/common/types';
-import { PythonV2DebugConfigurationProvider } from '../../../client/debugger';
 import { DebuggerTypeName } from '../../../client/debugger/Common/constants';
-import { DebugOptions, LaunchRequestArguments } from '../../../client/debugger/Common/Contracts';
-import { PythonLaunchDebugConfiguration } from '../../../client/debugger/configProviders/baseProvider';
-import { ConfigurationProviderUtils } from '../../../client/debugger/configProviders/configurationProviderUtils';
-import { IConfigurationProviderUtils } from '../../../client/debugger/configProviders/types';
+import { DebugOptions } from '../../../client/debugger/Common/Contracts';
+import { ConfigurationProviderUtils } from '../../../client/debugger/extension/configProviders/configurationProviderUtils';
+import { PythonV2DebugConfigurationProvider } from '../../../client/debugger/extension/configProviders/pythonV2Provider';
+import { IConfigurationProviderUtils } from '../../../client/debugger/extension/configProviders/types';
 import { IInterpreterHelper } from '../../../client/interpreter/contracts';
 import { IServiceContainer } from '../../../client/ioc/types';
 
@@ -289,7 +288,7 @@ suite('Debugging - Config Provider', () => {
         setupIoc(pythonPath);
         setupActiveEditor(pythonFile, PYTHON_LANGUAGE);
 
-        const debugConfig = await debugProvider.resolveDebugConfiguration!(workspaceFolder, { redirectOutput: false } as PythonLaunchDebugConfiguration<LaunchRequestArguments>);
+        const debugConfig = await debugProvider.resolveDebugConfiguration!(workspaceFolder, { redirectOutput: false } as any);
 
         expect(debugConfig).to.have.property('console', 'integratedTerminal');
         expect(debugConfig).to.have.property('stopOnEntry', false);
